test(docs): locate the theme module by code, not by text that mentions it
review-pr finding, and the only one that was a real hole rather than a
stale pointer. `rustSourceDefining` matched `pattern.test(f.text)` on the
raw file, which cannot tell a definition from a commented-out example or a
`#[cfg(test)]` mock. Review constructed the defeat: a block comment in
`win_app_mode.rs` holding two syntactically exact command signatures made
the locator resolve THERE, and -- the part that matters -- made the new
control spec pass while it did, because a commented-out `fn get_app_theme(`
satisfies `toContain` exactly as well as a real one. Seven other specs
still failed, so the suite did not go silent; the control written to rule
out "resolved to some unrelated file" did.

`RustSource` now carries `code` alongside `text` -- comments and
`#[cfg(test)]` modules stripped. The locator and the control read `code`;
the ten claim specs keep reading `text`, because those claims genuinely
are about attributes.

The two strippers were already in
startup-open-failure-wiring-claims.test.ts, carrying two hard-won
properties (a test could satisfy a production-routing claim; a byte-string
`b"{ not json"` defeats a naive brace counter). They move to
rust-sources.ts verbatim rather than being reimplemented, and that file's
own control on the stripper now covers the shared copy.

Measured both directions, because "renaming the real function goes red"
proves nothing on its own -- that is red either way. Running ONLY the
control spec: green at baseline, red under the hijack, and green under the
hijack with both halves of this fix reverted. That last run is what makes
the first two evidence.

Also repoints four more `lib.rs` pointers review found, in
win_app_mode.rs, useTauriTheme.svelte.ts (x2) and its test.
