+++ tests/docs/startup-open-failure-wiring-claims.test.ts
@@ -1,7 +1,13 @@
 import { readFileSync } from "node:fs";
 import { join } from "node:path";
 import { describe, expect, it } from "vitest";
-import { REPO_ROOT, RUST_SRC, rustSources } from "./rust-sources.js";
+import {
+  REPO_ROOT,
+  RUST_SRC,
+  rustSources,
+  stripRustComments,
+  stripRustTestModules,
+} from "./rust-sources.js";
 
 /**
  * Pins the parts of #1416's fix that **nothing else can fail on**.
@@ -38,110 +44,6 @@ import { REPO_ROOT, RUST_SRC, rustSources } from "./rust-sources.js";
 
 const LIB_RS = join(RUST_SRC, "lib.rs");
 const CLIENT_MAP = join(REPO_ROOT, "src", "client", "utils", "startup-rejection.ts");
-
-/**
- * Drop `#[cfg(test)] mod ... { ... }` blocks.
- *
- * **A test satisfied this file's production-routing claim.** The exclusion
- * check asks whether `CODE_UPDATE_MAY_NOT_HAVE_COMPLETED` is genuinely passed
- * to the pending-update surface -- and `pending_update_tests` calls
- * `surface_pending_update_hint_with(CODE_UPDATE_MAY_NOT_HAVE_COMPLETED, ...)`,
- * which matched. Rewriting the real call site to pass a literal instead left
- * this file green; found by mutation, not by reading. The hole predates the
- * Unit 11a extraction -- the test module was previously in `lib.rs` and equally
- * in scope -- but the widened scan is what made it worth closing here.
- *
- * Brace-counted rather than regex-bounded, and "strips test modules without
- * eating production code" is the control on it: a stripper that removed too
- * much would make every assertion below pass by finding nothing.
- *
- * **Braces are counted outside string and char literals**, because a naive
- * counter is defeated by text this very file already scans:
- * `pending_update.rs`'s test module contains `b"{ not json"`, an unbalanced
- * brace in a byte-string literal. A counter that sees it never reaches depth 0
- * and runs off the end of the input — today indistinguishable from a correct
- * match only because that test module is the last thing in the file. The next
- * constant appended after it would vanish from the parity scan silently. Found
- * by review, not by reading. Hence both halves: literals are skipped, and an
- * unbalanced block **throws** rather than truncating, because a stripper that
- * fails loud cannot hollow the scan it feeds.
- */
-function stripRustTestModules(src: string): string {
-  let out = "";
-  let i = 0;
-  for (;;) {
-    const at = src.indexOf("#[cfg(test)]", i);
-    if (at === -1) return out + src.slice(i);
-    const open = src.indexOf("{", at);
-    if (open === -1) return out + src.slice(i);
-    out += src.slice(i, at);
-    i = matchRustBrace(src, open) + 1;
-  }
-}
-
-/** Index of the `}` closing the `{` at `open`. Throws if the block never closes. */
-function matchRustBrace(src: string, open: number): number {
-  let depth = 0;
-  let j = open;
-  while (j < src.length) {
-    const c = src[j];
-    if (c === '"' || c === "'") {
-      const after = skipRustLiteral(src, j);
-      if (after > j) {
-        j = after;
-        continue;
-      }
-    }
-    if (c === "{") depth++;
-    else if (c === "}" && --depth === 0) return j;
-    j++;
-  }
-  throw new Error(
-    "unbalanced `#[cfg(test)]` block: the brace scanner ran off the end of the " +
-      "input. Truncating here would silently drop every Rust item after it from " +
-      "the wire-code scan, so this fails instead.",
-  );
-}
-
-/**
- * Index just past the string or char literal starting at `i`, or `i` if what is
- * there is not a literal (a lifetime `'a`, most commonly).
- *
- * Comments are already gone — `stripRustComments` runs first — so only literals
- * can hide a brace.
- */
-function skipRustLiteral(src: string, i: number): number {
-  if (src[i] === "'") {
-    // `'\n'` / `'{'` are literals; `'a` is a lifetime and must not be skipped.
-    if (src[i + 1] === "\\") {
-      const end = src.indexOf("'", i + 2);
-      return end === -1 ? i : end + 1;
-    }
-    return src[i + 2] === "'" ? i + 3 : i;
-  }
-  // Raw-ness is decided by looking back over `#`s to an `r` (`r"`, `r#"`, `br#"`).
-  let k = i - 1;
-  let hashes = 0;
-  while (src[k] === "#") {
-    hashes++;
-    k--;
-  }
-  if (src[k] === "r") {
-    const terminator = `"${"#".repeat(hashes)}`;
-    const end = src.indexOf(terminator, i + 1);
-    return end === -1 ? i : end + terminator.length;
-  }
-  for (let j = i + 1; j < src.length; j++) {
-    if (src[j] === "\\") j++;
-    else if (src[j] === '"') return j + 1;
-  }
-  return i;
-}
-
-/** Drop `//` and `/* *\/` comments, so prose about a call is not read as the call. */
-function stripRustComments(src: string): string {
-  return src.replace(/\/\*[\s\S]*?\*\//g, "").replace(/(^|[^:])\/\/.*$/gm, "$1");
-}
 
 function stripTsComments(src: string): string {
   return src.replace(/\/\*[\s\S]*?\*\//g, "").replace(/(^|[^:])\/\/.*$/gm, "$1");
